refactor:dto를 내부 클래스로 변경 및 deletedto 위치 변경
