+++ src/main/java/Bob_BE/domain/menu/service/MenuService.java
@@ -3,9 +3,9 @@ package Bob_BE.domain.menu.service;
 import Bob_BE.domain.menu.dto.request.MenuRequestDTO.MenuDeleteRequestDTO;
 import Bob_BE.domain.menu.dto.request.MenuRequestDTO.MenuUpdateRequestDTO;
 import Bob_BE.domain.menu.dto.response.MenuResponseDTO;
+import Bob_BE.domain.menu.dto.response.MenuResponseDTO.DeleteMenuResponseDTO;
 import Bob_BE.domain.menu.entity.Menu;
 import Bob_BE.domain.menu.repository.MenuRepository;
-import Bob_BE.domain.store.dto.response.StoreResponseDTO.DeleteMenuResponseDTO;
 import Bob_BE.global.response.code.resultCode.ErrorStatus;
 import Bob_BE.global.response.exception.handler.UserHandler;
 import java.util.List;
