+++ src/main/resources/application.yml
@@ -2,7 +2,7 @@ spring:
   datasource:
     url: jdbc:mysql://localhost:3306/bab
     username: root
-    password: khsoo2580
+    password: mqkrwnsgk0903!
     driver-class-name: com.mysql.cj.jdbc.Driver
   sql:
     init:
